requirements.txt
tensorflow==2.17.0
scikit-learn==1.5.2
keras==3.3.3
keras-facenet==0.3.2
Keras-Preprocessing==1.1.2
Flask==3.0.3
imutils==0.5.4
opencv-python==4.10.0.84
pillow==11.0.0